requirements.txt
pymodbus>=3.5.0
influxdb-client>=1.38.0
pyyaml>=6.0
asyncio>=3.4.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
